Move builder to inside Taggable, clean docs, default taggable.tag(Class)
